+++ main.h
@@ -10,9 +10,9 @@ int _print_binary(unsigned int num);
 int _print_number(int num);
 int _putchar(char c);
 int _putstr(char *str);
-int	_put_x(unsigned long nb);
-int	_put_X(unsigned long nb);
-int get_len(unsigned long nb);
+int	_put_x(unsigned int nb);
+int	_put_X(unsigned int nb);
+int get_len(unsigned int nb);
 int	_put_o(unsigned long nb);
 int _put_u(unsigned int num);
